storemgmt: Future proof the recognition of keytypes
This replaces key_is_rsa() with assign_key_data(), allowing it to add
whatever data is needed to the keymgmt data structure, while retaining
the return values that key_is_rsa() did.

The keymgmt data structure is extended with the keytype, a constant string
that doesn't need any extra management.

With this modification, load_another_private_key() should be able to
process any key that this provider supports, given a couple of added
lines in assign_key_data().