Add a GitHub workflow to build the provider
This removes the setting of CMAKE_INSTALL_PREFIX and OPENSSL_ROOT_DIR from
CMakeLists.txt, as these should be settable by the user as they see fit.
If you want to save settings anywhere, consider create a set of presets in
CMakeUserPresets.json.

The workflow ensures that the artifacts (cng_provider module and client app)
are statically linked with OpenSSL libraries.  This avoids a problem with
CMake's FindOpenSSL.cmake, which defines the found libraries with the type
UNKNOWN, which excludes declaring the actual OpenSSL DLLs, and thereby
doesn't install them in a way that works with a MODULE on Windows.