CMake files: A provider is a MODULE, not a SHARED library
On many platforms, the difference is very small.  However, they are there.
Most of all, this signals different usage expectations; an application is
expected to be linked with a SHARED library, but not with a MODULE.  A
MODULE is expected to be loaded dynamically by the application.

On Windows, this means that a MODULE isn't expected to come with an export
library (albeit many build systems still produce one, unnecessarily), and
isn't expected to be installed in the same directory as the application that
uses it.